# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: Codex Remote API Contract
  version: 0.1.0
  description: Control Plane-shaped contract shared by Web, Worker, Control Plane, and future iOS clients.
...
paths:
...
  /projects:
    get:
      operationId: listRemoteProjects
      parameters:
        - name: deviceId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Remote projects visible to the current actor.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RemoteProject"
  /conversations:
    get:
      operationId: listCodexConversations
      parameters:
        - name: deviceId
          in: query
          required: false
          schema:
            type: string
        - name: projectId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Codex conversations visible to the current actor.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CodexConversation"
...
  /v1/worker/capabilities:
    get:
      operationId: getWorkerCapabilities
      responses:
        "200":
          description: Read-only worker capabilities visible to the current actor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerCapabilities"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "408":
          $ref: "#/components/responses/RequestTimeoutError"
        "424":
          $ref: "#/components/responses/AppServerUnavailableError"
        "500":
          $ref: "#/components/responses/InternalWorkerError"
  /v1/worker/probe:
    get:
      operationId: getWorkerProbeSummary
      responses:
        "200":
          description: Latest worker-owned read-only probe summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerProbeSummary"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "408":
          $ref: "#/components/responses/RequestTimeoutError"
        "424":
          $ref: "#/components/responses/AppServerUnavailableError"
        "500":
          $ref: "#/components/responses/InternalWorkerError"
...
  /v1/projects:
    get:
      operationId: listControlPlaneProjects
      responses:
        "200":
          description: Projects visible through the Control Plane.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RemoteProject"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "424":
          $ref: "#/components/responses/DeviceUnavailableError"
        "500":
          $ref: "#/components/responses/InternalWorkerError"
...
  /v1/conversations:
    get:
      operationId: listWorkerConversations
      responses:
        "200":
          description: Worker conversation list visible to the current actor.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CodexConversation"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "408":
          $ref: "#/components/responses/RequestTimeoutError"
        "424":
          $ref: "#/components/responses/AppServerUnavailableError"
        "500":
          $ref: "#/components/responses/InternalWorkerError"
    post:
      operationId: startWorkerConversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartConversationInput"
      responses:
        "202":
          description: Start command accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandAccepted"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "408":
          $ref: "#/components/responses/RequestTimeoutError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "424":
          $ref: "#/components/responses/AppServerUnavailableError"
        "500":
          $ref: "#/components/responses/InternalWorkerError"
  /v1/conversations/{conversationId}/timeline:
    get:
      operationId: getConversationTimeline
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Timeline snapshot for a single conversation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationTimeline"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/ConversationNotFoundError"
        "408":
          $ref: "#/components/responses/RequestTimeoutError"
        "424":
          $ref: "#/components/responses/AppServerUnavailableError"
        "500":
          $ref: "#/components/responses/InternalWorkerError"
  /v1/conversations/{conversationId}:
    patch:
      operationId: renameWorkerConversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameConversationInput"
      responses:
        "200":
          description: Conversation renamed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenConversationResult"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/ConversationNotFoundError"
        "408":
          $ref: "#/components/responses/RequestTimeoutError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "424":
          $ref: "#/components/responses/AppServerUnavailableError"
        "500":
          $ref: "#/components/responses/InternalWorkerError"
  /v1/conversations/{conversationId}/open:
    post:
      operationId: openWorkerConversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationLifecycleInput"
      responses:
        "200":
          description: Conversation opened.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenConversationResult"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/ConversationNotFoundError"
        "408":
          $ref: "#/components/responses/RequestTimeoutError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "424":
          $ref: "#/components/responses/AppServerUnavailableError"
        "500":
          $ref: "#/components/responses/InternalWorkerError"
  /v1/conversations/{conversationId}/archive:
    post:
      operationId: archiveWorkerConversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationLifecycleInput"
      responses:
        "200":
          description: Conversation archived.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenConversationResult"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/ConversationNotFoundError"
        "408":
          $ref: "#/components/responses/RequestTimeoutError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "424":
          $ref: "#/components/responses/AppServerUnavailableError"
        "500":
          $ref: "#/components/responses/InternalWorkerError"
  /v1/conversations/{conversationId}/unarchive:
    post:
      operationId: unarchiveWorkerConversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationLifecycleInput"
      responses:
        "200":
          description: Conversation unarchived.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenConversationResult"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/ConversationNotFoundError"
        "408":
          $ref: "#/components/responses/RequestTimeoutError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "424":
          $ref: "#/components/responses/AppServerUnavailableError"
        "500":
          $ref: "#/components/responses/InternalWorkerError"
  /v1/conversations/{conversationId}/follow-up:
    post:
      operationId: followUpWorkerConversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FollowUpInput"
      responses:
        "202":
          description: Follow-up command accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandAccepted"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/ConversationNotFoundError"
        "408":
          $ref: "#/components/responses/RequestTimeoutError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "424":
          $ref: "#/components/responses/AppServerUnavailableError"
        "500":
          $ref: "#/components/responses/InternalWorkerError"
  /v1/conversations/{conversationId}/turns/{turnId}/interrupt:
    post:
      operationId: interruptWorkerTurn
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
        - name: turnId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterruptTurnInput"
      responses:
        "202":
          description: Interrupt command accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandAccepted"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/ConversationNotFoundError"
        "408":
          $ref: "#/components/responses/RequestTimeoutError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "424":
          $ref: "#/components/responses/AppServerUnavailableError"
        "500":
          $ref: "#/components/responses/InternalWorkerError"
  /v1/conversations/{conversationId}/turns/{turnId}/steer:
    post:
      operationId: steerWorkerTurn
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
        - name: turnId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SteerTurnInput"
      responses:
        "202":
          description: Steer command accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandAccepted"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/ConversationNotFoundError"
        "408":
          $ref: "#/components/responses/RequestTimeoutError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "424":
          $ref: "#/components/responses/AppServerUnavailableError"
        "500":
          $ref: "#/components/responses/InternalWorkerError"
  /v1/conversations/{conversationId}/approvals:
    get:
      operationId: listPendingApprovals
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Process-local pending approval snapshot for a conversation.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PendingApproval"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/ConversationNotFoundError"
        "408":
          $ref: "#/components/responses/RequestTimeoutError"
        "424":
          $ref: "#/components/responses/AppServerUnavailableError"
        "500":
          $ref: "#/components/responses/InternalWorkerError"
  /v1/conversations/{conversationId}/approvals/{approvalRequestId}/decision:
    post:
      operationId: decidePendingApproval
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
        - name: approvalRequestId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionInput"
      responses:
        "202":
          description: Approval decision command accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandAccepted"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/ConversationNotFoundError"
        "408":
          $ref: "#/components/responses/RequestTimeoutError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "424":
          $ref: "#/components/responses/AppServerUnavailableError"
        "500":
          $ref: "#/components/responses/InternalWorkerError"
...
components:
  responses:
    BadRequestError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    UnauthorizedError:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    ForbiddenError:
      description: Caller is not allowed to access the requested worker resource.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    ConversationNotFoundError:
      description: The requested conversation timeline could not be found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    RequestTimeoutError:
      description: Worker request timed out before a stable response was produced.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    ConflictError:
      description: Request conflicts with an existing write command.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    AppServerUnavailableError:
      description: Worker could not reach the local Codex app-server dependency.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    DeviceUnavailableError:
      description: Control Plane could not reach the configured device Worker.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    InternalWorkerError:
      description: Worker hit an unexpected internal failure.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    ConversationStatus:
      type: string
      enum:
        - running
        - waiting
        - done
        - failed
        - unknown
...
    ProbeMode:
      type: string
      enum:
        - readOnly
        - full
...
    WorkerCapabilities:
      type: object
      additionalProperties: false
      required:
        - deviceId
        - canReadProjects
        - canReadConversations
        - canReadTimeline
        - canRunReadOnlyProbe
        - appServerTransport
        - supportedSourceKinds
      properties:
        deviceId:
          type: string
        canReadProjects:
          type: boolean
        canReadConversations:
          type: boolean
        canReadTimeline:
          type: boolean
        canRunReadOnlyProbe:
          type: boolean
        appServerTransport:
          $ref: "#/components/schemas/AppServerTransport"
        supportedSourceKinds:
          type: array
          items:
            type: string
...
    ConversationTimeline:
      type: object
      additionalProperties: false
      required:
        - deviceId
        - conversationId
        - readStartedAt
        - readCompletedAt
        - snapshotRevision
        - runtimeStatus
        - latestTurnStatus
        - turns
      properties:
        deviceId:
          type: string
        conversationId:
          type: string
        projectId:
          type: string
        readStartedAt:
          type: string
        readCompletedAt:
          type: string
        snapshotRevision:
          type: string
        runtimeStatus:
          $ref: "#/components/schemas/ConversationRuntimeStatus"
        latestTurnStatus:
          $ref: "#/components/schemas/LatestTurnStatus"
        loaded:
          type: boolean
        live:
          type: boolean
        archived:
          type: boolean
        turns:
          type: array
          items:
            $ref: "#/components/schemas/ConversationTimelineTurn"
        events:
          type: array
          items:
            $ref: "#/components/schemas/ConversationWorkbenchEvent"
...
    WorkerProbeSummary:
      type: object
      additionalProperties: false
      required:
        - schemaVersion
        - startedAt
        - completedAt
        - ok
        - mode
        - deviceId
        - codexVersion
        - appServer
        - checks
      properties:
        schemaVersion:
          type: number
        startedAt:
          type: string
        completedAt:
          type: string
        ok:
          type: boolean
        mode:
          $ref: "#/components/schemas/ProbeMode"
        deviceId:
          type: string
        codexVersion:
          type:
            - string
            - "null"
        appServer:
          type: object
          additionalProperties: false
          required:
            - transport
            - startedByWorker
            - readyz
          properties:
            transport:
              $ref: "#/components/schemas/AppServerTransport"
            startedByWorker:
              type: boolean
            readyz:
              type: boolean
        checks:
          type: array
          items:
            $ref: "#/components/schemas/ProbeCheckResult"
...
    RemoteProject:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - deviceId
        - path
        - branch
        - hasChanges
        - pinned
      properties:
        id:
          type: string
        name:
          type: string
        deviceId:
          type: string
        path:
          type: string
        branch:
          type: string
        hasChanges:
          type: boolean
        pinned:
          type: boolean
        expanded:
          type: boolean
    CodexConversation:
      type: object
      additionalProperties: false
      required:
        - id
        - title
        - deviceId
        - projectName
        - status
        - updatedAt
        - summary
        - sandbox
        - approval
      properties:
        id:
          type: string
        title:
          type: string
        deviceId:
          type: string
        projectId:
          type: string
        projectName:
          type: string
        status:
          $ref: "#/components/schemas/ConversationStatus"
        updatedAt:
          type: string
        summary:
          type: string
        sandbox:
          type: string
        approval:
          type: string
        archived:
          type: boolean
        loaded:
          type: boolean
        live:
          type: boolean
        pinned:
          type: boolean
...
    StartConversationInput:
      type: object
      additionalProperties: false
      required:
        - projectId
        - message
        - clientRequestId
      properties:
        projectId:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
          maxLength: 20000
        clientRequestId:
          type: string
          minLength: 1
          maxLength: 128
...
    RenameConversationInput:
      type: object
      additionalProperties: false
      required:
        - title
        - clientRequestId
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
        clientRequestId:
          type: string
          minLength: 1
          maxLength: 128
    OpenConversationResult:
      type: object
      additionalProperties: false
      required:
        - conversation
        - timeline
      properties:
        conversation:
          $ref: "#/components/schemas/CodexConversation"
        timeline:
          $ref: "#/components/schemas/ConversationTimeline"
    CommandAccepted:
      type: object
      additionalProperties: false
      required:
        - id
        - status
        - conversationId
        - turnId
        - acceptedAt
      properties:
        id:
          type: string
        status:
          type: string
          enum:
            - accepted
        conversationId:
          type: string
        turnId:
          type:
            - string
            - "null"
        acceptedAt:
          type: string
          format: date-time
...